House Cleaning
- moved pipeCase()() into Pipe BotAction's from Utilities
- moved tests accordingly
- tweaked pipeCase()() for callback's to set function as value instead
  of evaluated value (always true) - prob more interested in the index
