+++ src/botmation/types/abort-line-signal.ts
