+++ src/botmation/types/index.ts
@@ -3,4 +3,4 @@ export * from './bot-indexed-db-inject'
 export * from './database'
 export * from './objects'
 export * from './pipe-value'
-export * from './abort-signal'
+export * from './abort-line-signal'
